sliderCards/totalConsumptionSeperated BarGraph done

+++ app/src/main/java/com/deksi/bpsfmmobileapp/home/model/Dashboard.kt
@@ -160,7 +160,7 @@ data class gridTotalConsumption(
     @SerializedName("percentageDiff") var percentageDiff: Int? = null,
     @SerializedName("name") var name: String? = null,
     @SerializedName("cost") var cost: Int? = null,
-    @SerializedName("energy") var energy: Double? = null,
+    @SerializedName("energy") var energy: Float? = null,
     @SerializedName("ts") var ts: String? = null
 
 )
